refactor(folder-structure): fix folder path for internal components
