+++ app/(dashboard)/dashboard/_components/filter-options.tsx
@@ -8,7 +8,7 @@ import { useCallback, useState } from "react"
 import { OptionTab } from "./option-tab";
 
 
-export const QueryOptions = ({ itemStatus }: { itemStatus: string }) => {
+export const FilterOptions = ({ itemStatus }: { itemStatus: string }) => {
   // STATE
   const [status, setStatus] = useState(itemStatus || 'ALL');
   const searchParams = useSearchParams();
