+++ app/(dashboard)/dashboard/page.tsx
@@ -1,10 +1,10 @@
 // DATA
 import { getCollectionData } from '@/data/firestore';
 // COMPONENTS
-import { TableContainer } from '../_components/table-container';
-import { ContainerOptions } from '../_components/container-options';
-import { GridContainer } from '../_components/grid-container';
-import { QueryOptions } from '../_components/query-options';
+import { TableContainer } from './_components/table-container';
+import { ViewOptions } from './_components/container-options';
+import { GridContainer } from './_components/grid-container';
+import { FilterOptions } from './_components/filter-options';
 
 
 export default async function DashboardPage({ searchParams }: { searchParams: any }) {
@@ -31,8 +31,8 @@ export default async function DashboardPage({ searchParams }: { searchParams: an
   return (
     <>
       <section className="flex flex-row justify-between items-stetch flex-wrap">
-        <QueryOptions itemStatus={searchParams.status} />
-        <ContainerOptions displayType={searchParams.display} />
+        <FilterOptions itemStatus={searchParams.status} />
+        <ViewOptions displayType={searchParams.display} />
       </section>
       {renderContainer(searchParams.display)}
     </>
